[RMIS-57] refactor : modify form page 코드 분할
