+++ src/components/admin/modify-form/form-questions/checkbox-or-radion-question.tsx
@@ -1,0 +1,12 @@
+import { Input } from '@/components/common/form/input';
+
+export function CheckboxOrRadioQuestion() {
+  return (
+    <>
+      <Input type='text' />
+      <Input type='text' />
+      <Input type='text' />
+      <button>+</button>
+    </>
+  );
+}
