+++ src/components/admin/modify-form/form-questions/number-question.tsx
@@ -1,0 +1,5 @@
+import { Input } from '@/components/common/form/input';
+
+export function NumberQuestion() {
+  return <Input type='number' />;
+}
