+++ src/components/admin/modify-form/form-questions/text-question.tsx
@@ -1,0 +1,5 @@
+import { Input } from '@/components/common/form/input';
+
+export function TextQuestion() {
+  return <Input readOnly />;
+}
